% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gitignore.R
\name{write_gitignore}
\alias{write_gitignore}
\title{Make a gitignore file}
\usage{
write_gitignore(
  ignore = ignore_items,
  file = ".gitignore",
  add = file.exists(file)
)
}
\description{
A function to write a .gitignore file with common ignore items
}
\details{
To see the vector of default things added to .gitignore, call voovarb:::ignore_items
}
\author{
Myfanwy Johnston

Matt Espe
}
